Release v0.6.2
Reconstructed from the shipped release archive. No version control
existed for this release; the archive is the only surviving record.

X-Provenance: e107_v0.602.tar.gz
X-Provenance-SHA256: a42926c3636358e3ad950a4b54088614fc09a869187e8b815d8f1bf1459fb422
X-Provenance-Files: 934
X-Author-Basis: source header in e107_v0.602.tar.gz
X-Date-Basis: author=max member mtime; committer=SourceForge publication
X-Release-Kind: release
X-SourceForge-Folder: e107/e107 v0.602

+++ e107_admin/sql/core_sql.php
@@ -326,7 +326,7 @@ CREATE TABLE online (
 # --------------------------------------------------------
 
 #
-# Table structure for table `poll`
+# Table structure for table `plugin`
 #
 
 CREATE TABLE plugin (
@@ -347,7 +347,7 @@ CREATE TABLE poll (
   poll_id int(10) unsigned NOT NULL auto_increment,
   poll_datestamp int(10) unsigned NOT NULL default '0',
   poll_end_datestamp int(10) unsigned NOT NULL default '0',
-  poll_admin_id tinyint(3) unsigned NOT NULL default '0',
+  poll_admin_id int(10) unsigned NOT NULL default '0',
   poll_title varchar(250) NOT NULL default '',
   poll_option_1 varchar(250) NOT NULL default '',
   poll_option_2 varchar(250) NOT NULL default '',
